dorra commit

+++ crmproject-ejb/src/main/java/crm/impl/prospecting/PointOfSaleImpl.java
@@ -6,8 +6,11 @@ import javax.ejb.LocalBean;
 import javax.ejb.Stateless;
 import javax.persistence.*;
 import javax.persistence.Query;
+
+import crm.entities.Roles;
 import crm.entities.prospecting.*;
 import crm.interfaces.prospecting.*;
+import crm.utils.UserSession;
 
 @Stateless
 @LocalBean
@@ -18,41 +21,47 @@ public class PointOfSaleImpl implements IPointOfSaleLocal, IPointOfSaleRemote {
 
 	@Override
 	public List<PointOfSale> allPointOfSale() {
-		Query q = em.createQuery("SELECT p.id, p.namePS, pp.latitude , pp.longitude FROM PointOfSale p JOIN p.location pp");
+		TypedQuery<PointOfSale> q = em.createQuery("SELECT c FROM PointOfSale c", PointOfSale.class);
 		return (List<PointOfSale>) q.getResultList();
 	}
 
 	@Override
 	public List<PointOfSale> searchForPointOfSale(String POSName) {
-		Query q = em.createQuery("SELECT p.id, p.namePS, pp.latitude , pp.longitude FROM PointOfSale p JOIN p.location pp where p.namePS = :POSName");
+		Query q = em.createQuery("SELECT p FROM PointOfSale p JOIN p.location pp where p.namePS = :POSName");
 		q.setParameter("POSName", POSName);
 		return (List<PointOfSale> ) q.getResultList(); 
 	}
 
 	
 	@Override
-	public void addPointOfSale(String name, float latitude, float longitude) {
-		Location location = new Location(latitude, longitude); 
-		PointOfSale pos = new PointOfSale(name, location);
-		em.persist(pos);
+	public boolean addPointOfSale(String name, float latitude, float longitude) {
+		if (UserSession.getInstance().getRole()== Roles.ADMIN)
+		{
+			Location location = new Location(latitude, longitude); 
+			PointOfSale pos = new PointOfSale(name, location);
+			em.persist(pos);
+			return true ; 
+			
+		}
+		return false ; 
+		
 		
 	}
 
 	@Override
-	public boolean deletePointOfSale(int id) {
-		PointOfSale pos = em.find(PointOfSale.class, id); 
-		if (pos != null )
+	public int deletePointOfSale(int id) {
+		PointOfSale pos = em.find(PointOfSale.class, id);
+		if (UserSession.getInstance().getRole()== Roles.ADMIN)
 		{
-			em.remove(pos);
-			//Query q = em.createQuery("DELETE FROM PointOfSale p WHERE p.id = :id");
-	       // q.setParameter("id", id);
-	       // q.executeUpdate();
-	        return true ; 
+			if (pos != null )
+			{
+				em.remove(pos);
+		        return 1 ; //deleted
+				
+			}else return -1; //pos not found
 			
 		}
-		
-		
-		return false; 
+		return 0; //not an admin
 	}
 
 	@Override
@@ -60,16 +69,19 @@ public class PointOfSaleImpl implements IPointOfSaleLocal, IPointOfSaleRemote {
 	
 		Location location = new Location(latitude, longitude); 
 		PointOfSale p = em.find(PointOfSale.class,id); 
+		if (UserSession.getInstance().getRole()== Roles.ADMIN)
+		{
 		if (p!= null)
 		{
 			p.setNamePS(name);
 			p.setLocation(location);
 			em.merge(p); 
-			return p.getId(); 
+			return 1; 
 			
 		}
-		return 0; 
-		
+		else return -1; //not found 
+		}
+		return 0; //not an admin
 		
 	}
